refactored and added a folder for GameScreens
I could convert JavaFXScreen into a concrete class and forgo abstraction in JavaFXScreens
